Update main.py
In the 'Seller  per product' model I have moved 'unique is true' to the product (=product_id)because not every seller could only be linked to 1 product, but each product may only be linked to 1 seller. 
The "search and the list functions" return the value instead of printing.
I've added "try excepts" to the functions to return a message.
And added thr condition that there must be enough in stock for a purchase.